+++ src/moduleStore/index.js
@@ -257,7 +257,14 @@ export default {
         
         tags: m.tags,
         github: m.github,
-        images: m.images,
+        images: m.images.map((x) => {
+          if (x.startsWith('/')) {
+            const baseUrl = m.repo.split('/').slice(0, -1).join('/');
+            x = baseUrl + x;
+          }
+
+          return x;
+        }),
         lastUpdated: m.lastUpdated,
 
         buttonType: goosemodScope.modules[m.name] || goosemodScope.disabledModules[m.name] ? 'danger' : 'brand',