+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
   - ### Store
     - Fixed id cache not being invalidated every day
+    - Added localised path support for repo images, like using /images/file.png
 
   - ### i18n
     - Fixed getting Discord's values not working, caused settings and more to not appear correctly